Move GdgListFragment into common packge
